+++ src/main/java/uk/gov/hmcts/divorce/document/content/JudicialSeparationCoRefusalTemplateContent.java
@@ -5,7 +5,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.TemplateContent;
 
 import java.time.Clock;
 import java.time.LocalDate;
