+++ src/main/java/uk/gov/hmcts/divorce/legaladvisor/notification/LegalAdvisorRejectedDecisionNotification.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.ConditionalOrderRefusalDocumentPack;
+import uk.gov.hmcts.divorce.document.print.documentpack.ConditionalOrderRefusalDocumentPack;
 import uk.gov.hmcts.divorce.legaladvisor.service.printer.LetterPrinter;
 import uk.gov.hmcts.divorce.notification.ApplicantNotification;
 import uk.gov.hmcts.divorce.notification.CommonContent;
