+++ src/test/java/uk/gov/hmcts/divorce/legaladvisor/notification/LegalAdvisorMoreInfoDecisionNotificationTest.java
@@ -8,8 +8,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.divorcecase.model.Solicitor;
 import uk.gov.hmcts.divorce.document.model.DocumentType;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.ConditionalOrderRefusalDocumentPack;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.DocumentPackInfo;
+import uk.gov.hmcts.divorce.document.print.documentpack.ConditionalOrderRefusalDocumentPack;
+import uk.gov.hmcts.divorce.document.print.documentpack.DocumentPackInfo;
 import uk.gov.hmcts.divorce.legaladvisor.service.printer.LetterPrinter;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.NotificationService;
