+++ src/test/java/uk/gov/hmcts/divorce/legaladvisor/notification/LegalAdvisorRejectedDecisionNotificationTest.java
@@ -10,8 +10,8 @@ import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.divorcecase.model.Solicitor;
 import uk.gov.hmcts.divorce.document.model.DocumentType;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.ConditionalOrderRefusalDocumentPack;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.DocumentPackInfo;
+import uk.gov.hmcts.divorce.document.print.documentpack.ConditionalOrderRefusalDocumentPack;
+import uk.gov.hmcts.divorce.document.print.documentpack.DocumentPackInfo;
 import uk.gov.hmcts.divorce.legaladvisor.service.printer.LetterPrinter;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.NotificationService;
