+++ src/test/java/uk/gov/hmcts/divorce/legaladvisor/service/printer/LetterPrinterTest.java
@@ -15,12 +15,10 @@ import uk.gov.hmcts.divorce.document.model.DocumentType;
 import uk.gov.hmcts.divorce.document.print.BulkPrintService;
 import uk.gov.hmcts.divorce.document.print.model.Letter;
 import uk.gov.hmcts.divorce.document.print.model.Print;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.CoRefusalDocumentPack;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.DocumentGenerator;
-import uk.gov.hmcts.divorce.legaladvisor.service.conditionalorder.DocumentPackInfo;
+import uk.gov.hmcts.divorce.document.DocumentGenerator;
+import uk.gov.hmcts.divorce.document.print.documentpack.DocumentPackInfo;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
